requirements-webapp.txt
# Core AI/ML dependencies
torch==2.2.0
torchaudio==2.2.0
numpy==1.26.4  # Using v1.x as PyTorch 2.2.0 is not compatible with NumPy 2.x
scipy>=1.7.0,<2.0.0
pandas>=1.3.0,<2.3.0

# Audio processing
soundfile==0.13.1
ffmpeg-python>=0.2.0

# Whisper for transcription
openai-whisper==20240930
tiktoken==0.5.2

# PyAnnote for diarization
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyannote.database==5.1.3

# Document processing
python-docx==1.1.2

# Progress visualizations
tqdm==4.67.1

# PDF generation and templating
jinja2>=3.1.0,<3.2.0
markdown>=3.8.0
pyyaml>=6.0

# API and web interaction
requests>=2.25.0

# Flask and web framework
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
wtforms==3.1.1

# Background task processing
celery==5.3.4
redis==5.0.1

# Production server
gunicorn==21.2.0

# Environment and configuration
python-dotenv==1.0.0

# AI/ML additional dependencies
aiohappyeyeballs>=2.6.0
aiohttp>=3.8.0
huggingface-hub>=0.15.0
more-itertools>=8.0.0
regex>=2022.0.0

# Database drivers (optional for PostgreSQL in production)
# psycopg2-binary==2.9.7

# Development dependencies (optional)
# pytest==7.4.0
# pytest-flask==1.2.0
# flask-testing==0.8.1